JAMES-1877 Move class RemoteDeliverySocketFactory to a remoteDelivery dedicated package
